encryption function is now completed
This mark the final step I needed to start creating my backup.
now I just need to implement the code that would be required to reverse all of the changes back to the source directory after all this wouldn't be a good backup solution if I couldn't restore my backups.